resolved issue with updateGame method
the updateGame() method had an issue where the currentState was changing
to GameState.Nought_won when either player won. This was fixed by adding
break; commands to all cases in the switch statement. The break command
prevents the other cases from being executed. The correct win result is
now displayed.